Remove deprecated CLI commands
