+++ internal/cli/builds/builds_latest_test.go
@@ -1,196 +1,10 @@
 package builds
 
 import (
-	"context"
-	"errors"
-	"flag"
-	"path/filepath"
-	"strings"
 	"testing"
 
 	"github.com/rudrankriyam/App-Store-Connect-CLI/internal/asc"
 )
-
-func isolateBuildsAuthEnv(t *testing.T) {
-	t.Helper()
-
-	// Keep tests hermetic: avoid loading host keychain/config/env credentials.
-	t.Setenv("ASC_BYPASS_KEYCHAIN", "1")
-	t.Setenv("ASC_PROFILE", "")
-	t.Setenv("ASC_CONFIG_PATH", filepath.Join(t.TempDir(), "nonexistent.json"))
-	t.Setenv("ASC_KEY_ID", "")
-	t.Setenv("ASC_ISSUER_ID", "")
-	t.Setenv("ASC_PRIVATE_KEY_PATH", "")
-	t.Setenv("ASC_PRIVATE_KEY", "")
-	t.Setenv("ASC_PRIVATE_KEY_B64", "")
-	t.Setenv("ASC_STRICT_AUTH", "")
-}
-
-func TestBuildsLatestCommand_MissingApp(t *testing.T) {
-	isolateBuildsAuthEnv(t)
-
-	// Clear env var to ensure --app is required
-	t.Setenv("ASC_APP_ID", "")
-
-	cmd := BuildsLatestCommand()
-
-	err := cmd.Exec(context.Background(), []string{})
-	if !errors.Is(err, flag.ErrHelp) {
-		t.Errorf("expected flag.ErrHelp when --app is missing, got %v", err)
-	}
-}
-
-func TestBuildsLatestCommand_InvalidPlatform(t *testing.T) {
-	isolateBuildsAuthEnv(t)
-
-	cmd := BuildsLatestCommand()
-
-	// Parse flags first
-	if err := cmd.FlagSet.Parse([]string{"--app", "123", "--platform", "INVALID"}); err != nil {
-		t.Fatalf("failed to parse flags: %v", err)
-	}
-
-	err := cmd.Exec(context.Background(), []string{})
-	if !errors.Is(err, flag.ErrHelp) {
-		t.Errorf("expected flag.ErrHelp for invalid platform, got %v", err)
-	}
-}
-
-func TestBuildsLatestCommand_ValidPlatforms(t *testing.T) {
-	isolateBuildsAuthEnv(t)
-
-	validPlatforms := []string{"IOS", "MAC_OS", "TV_OS", "VISION_OS", "ios", "mac_os"}
-
-	for _, platform := range validPlatforms {
-		t.Run(platform, func(t *testing.T) {
-			isolateBuildsAuthEnv(t)
-
-			cmd := BuildsLatestCommand()
-
-			// Parse flags - this should not error for valid platforms
-			if err := cmd.FlagSet.Parse([]string{"--app", "123", "--platform", platform}); err != nil {
-				t.Fatalf("failed to parse flags: %v", err)
-			}
-
-			// The command will fail because there's no real client, but it should get past validation
-			err := cmd.Exec(context.Background(), []string{})
-
-			// Should not be flag.ErrHelp for valid platforms (will fail later due to no auth)
-			if errors.Is(err, flag.ErrHelp) {
-				t.Errorf("platform %s should be valid but got flag.ErrHelp", platform)
-			}
-		})
-	}
-}
-
-func TestBuildsLatestCommand_InvalidInitialBuildNumber(t *testing.T) {
-	isolateBuildsAuthEnv(t)
-
-	cmd := BuildsLatestCommand()
-
-	if err := cmd.FlagSet.Parse([]string{"--app", "123", "--next", "--initial-build-number", "0"}); err != nil {
-		t.Fatalf("failed to parse flags: %v", err)
-	}
-
-	err := cmd.Exec(context.Background(), []string{})
-	if !errors.Is(err, flag.ErrHelp) {
-		t.Errorf("expected flag.ErrHelp for invalid initial build number, got %v", err)
-	}
-}
-
-func TestBuildsLatestCommand_NotExpiredAndExcludeExpiredTogether(t *testing.T) {
-	isolateBuildsAuthEnv(t)
-
-	cmd := BuildsLatestCommand()
-
-	if err := cmd.FlagSet.Parse([]string{"--app", "123", "--exclude-expired", "--not-expired"}); err != nil {
-		t.Fatalf("failed to parse flags: %v", err)
-	}
-
-	err := cmd.Exec(context.Background(), []string{})
-	if errors.Is(err, flag.ErrHelp) {
-		t.Errorf("expected non-help execution path when both alias flags are set, got %v", err)
-	}
-}
-
-func TestBuildsLatestCommand_FlagDefinitions(t *testing.T) {
-	isolateBuildsAuthEnv(t)
-
-	cmd := BuildsLatestCommand()
-
-	// Verify all expected flags exist
-	expectedFlags := []string{"app", "version", "platform", "processing-state", "output", "pretty", "next", "initial-build-number", "exclude-expired", "not-expired"}
-	for _, name := range expectedFlags {
-		f := cmd.FlagSet.Lookup(name)
-		if f == nil {
-			t.Errorf("expected flag --%s to be defined", name)
-		}
-	}
-
-	// Verify default values
-	if f := cmd.FlagSet.Lookup("output"); f != nil && f.DefValue != "json" {
-		t.Errorf("expected --output default to be 'json', got %q", f.DefValue)
-	}
-	if f := cmd.FlagSet.Lookup("pretty"); f != nil && f.DefValue != "false" {
-		t.Errorf("expected --pretty default to be 'false', got %q", f.DefValue)
-	}
-	if f := cmd.FlagSet.Lookup("next"); f != nil && f.DefValue != "false" {
-		t.Errorf("expected --next default to be 'false', got %q", f.DefValue)
-	}
-	if f := cmd.FlagSet.Lookup("initial-build-number"); f != nil && f.DefValue != "1" {
-		t.Errorf("expected --initial-build-number default to be '1', got %q", f.DefValue)
-	}
-	if f := cmd.FlagSet.Lookup("exclude-expired"); f != nil && f.DefValue != "false" {
-		t.Errorf("expected --exclude-expired default to be 'false', got %q", f.DefValue)
-	}
-	if f := cmd.FlagSet.Lookup("not-expired"); f != nil && f.DefValue != "false" {
-		t.Errorf("expected --not-expired default to be 'false', got %q", f.DefValue)
-	}
-}
-
-func TestBuildsLatestCommand_ProcessingStateFlagDescription(t *testing.T) {
-	cmd := BuildsLatestCommand()
-
-	processingStateFlag := cmd.FlagSet.Lookup("processing-state")
-	if processingStateFlag == nil {
-		t.Fatal("expected --processing-state flag to be defined")
-		return
-	}
-	if !strings.Contains(processingStateFlag.Usage, "VALID") || !strings.Contains(processingStateFlag.Usage, "all") {
-		t.Fatalf("expected --processing-state usage to mention supported values, got %q", processingStateFlag.Usage)
-	}
-}
-
-func TestBuildsLatestCommand_HelpMentionsCanonicalReplacements(t *testing.T) {
-	cmd := BuildsLatestCommand()
-	if !strings.Contains(cmd.LongHelp, "asc builds info --latest") {
-		t.Fatalf("expected help text to mention asc builds info --latest")
-	}
-	if !strings.Contains(cmd.LongHelp, "asc builds next-build-number") {
-		t.Fatalf("expected help text to mention asc builds next-build-number")
-	}
-}
-
-func TestBuildsLatestCommand_UsesAppIDEnv(t *testing.T) {
-	isolateBuildsAuthEnv(t)
-
-	// Set env var
-	t.Setenv("ASC_APP_ID", "env-app-id")
-
-	cmd := BuildsLatestCommand()
-
-	// Don't pass --app flag
-	if err := cmd.FlagSet.Parse([]string{}); err != nil {
-		t.Fatalf("failed to parse flags: %v", err)
-	}
-
-	err := cmd.Exec(context.Background(), []string{})
-
-	// Should not be flag.ErrHelp since env var provides the app ID
-	if errors.Is(err, flag.ErrHelp) {
-		t.Errorf("should use ASC_APP_ID env var but got flag.ErrHelp")
-	}
-}
 
 // TestSelectNewestBuild verifies that the multi-preReleaseVersion selection
 // logic correctly picks the build with the newest uploadedDate.
