+++ internal/cli/cmdtest/builds_next_number_test.go
@@ -506,143 +506,6 @@ func TestBuildsNextBuildNumberVersionAndPlatformPaginatesPastNearMatches(t *test
 	}
 }
 
-func TestBuildsLatestAliasWarnsAndMatchesCanonicalNextBuildNumberOutput(t *testing.T) {
-	setupAuth(t)
-	t.Setenv("ASC_CONFIG_PATH", filepath.Join(t.TempDir(), "nonexistent.json"))
-
-	originalTransport := http.DefaultTransport
-	t.Cleanup(func() {
-		http.DefaultTransport = originalTransport
-	})
-
-	http.DefaultTransport = roundTripFunc(func(req *http.Request) (*http.Response, error) {
-		switch {
-		case req.Method == http.MethodGet && req.URL.Path == "/v1/builds":
-			body := `{
-				"data":[{"type":"builds","id":"build-1","attributes":{"version":"9","uploadedDate":"2026-02-01T00:00:00Z"}}]
-			}`
-			return &http.Response{
-				StatusCode: http.StatusOK,
-				Body:       io.NopCloser(strings.NewReader(body)),
-				Header:     http.Header{"Content-Type": []string{"application/json"}},
-			}, nil
-		case req.Method == http.MethodGet && req.URL.Path == "/v1/apps/100000001/buildUploads":
-			body := `{"data":[],"links":{"next":""}}`
-			return &http.Response{
-				StatusCode: http.StatusOK,
-				Body:       io.NopCloser(strings.NewReader(body)),
-				Header:     http.Header{"Content-Type": []string{"application/json"}},
-			}, nil
-		default:
-			t.Fatalf("unexpected request: %s %s", req.Method, req.URL.String())
-			return nil, nil
-		}
-	})
-
-	run := func(args []string) (string, string) {
-		root := RootCommand("1.2.3")
-		root.FlagSet.SetOutput(io.Discard)
-
-		return captureOutput(t, func() {
-			if err := root.Parse(args); err != nil {
-				t.Fatalf("parse error: %v", err)
-			}
-			if err := root.Run(context.Background()); err != nil {
-				t.Fatalf("run error: %v", err)
-			}
-		})
-	}
-
-	canonicalStdout, canonicalStderr := run([]string{"builds", "next-build-number", "--app", "100000001"})
-	aliasStdout, aliasStderr := run([]string{"builds", "latest", "--app", "100000001", "--next"})
-
-	if canonicalStderr != "" {
-		t.Fatalf("expected canonical command to avoid warnings, got %q", canonicalStderr)
-	}
-	requireStderrContainsWarning(t, aliasStderr, "Warning: `asc builds latest --next` is deprecated. Use `asc builds next-build-number`.")
-	assertOnlyDeprecatedCommandWarnings(t, aliasStderr)
-	if canonicalStdout != aliasStdout {
-		t.Fatalf("expected canonical and alias output to match, canonical=%q alias=%q", canonicalStdout, aliasStdout)
-	}
-}
-
-func TestBuildsLatestAliasWarnsAndMatchesCanonicalInfoLatestOutput(t *testing.T) {
-	setupAuth(t)
-	t.Setenv("ASC_CONFIG_PATH", filepath.Join(t.TempDir(), "nonexistent.json"))
-
-	originalTransport := http.DefaultTransport
-	t.Cleanup(func() {
-		http.DefaultTransport = originalTransport
-	})
-
-	http.DefaultTransport = roundTripFunc(func(req *http.Request) (*http.Response, error) {
-		switch req.URL.Path {
-		case "/v1/builds":
-			body := `{"data":[{"type":"builds","id":"build-latest","attributes":{"version":"42","uploadedDate":"2026-02-01T00:00:00Z"}}]}`
-			return &http.Response{
-				StatusCode: http.StatusOK,
-				Body:       io.NopCloser(strings.NewReader(body)),
-				Header:     http.Header{"Content-Type": []string{"application/json"}},
-			}, nil
-		case "/v1/builds/build-latest/preReleaseVersion":
-			body := `{"data":{"type":"preReleaseVersions","id":"prv-1"}}`
-			return &http.Response{
-				StatusCode: http.StatusOK,
-				Body:       io.NopCloser(strings.NewReader(body)),
-				Header:     http.Header{"Content-Type": []string{"application/json"}},
-			}, nil
-		default:
-			t.Fatalf("unexpected request: %s %s", req.Method, req.URL.String())
-			return nil, nil
-		}
-	})
-
-	run := func(args []string) (string, string) {
-		root := RootCommand("1.2.3")
-		root.FlagSet.SetOutput(io.Discard)
-
-		return captureOutput(t, func() {
-			if err := root.Parse(args); err != nil {
-				t.Fatalf("parse error: %v", err)
-			}
-			if err := root.Run(context.Background()); err != nil {
-				t.Fatalf("run error: %v", err)
-			}
-		})
-	}
-
-	canonicalStdout, canonicalStderr := run([]string{"builds", "info", "--app", "100000001", "--latest"})
-	aliasStdout, aliasStderr := run([]string{"builds", "latest", "--app", "100000001"})
-
-	if canonicalStderr != "" {
-		t.Fatalf("expected canonical command to avoid warnings, got %q", canonicalStderr)
-	}
-	requireStderrContainsWarning(t, aliasStderr, "Warning: `asc builds latest` is deprecated. Use `asc builds info --latest`.")
-	assertOnlyDeprecatedCommandWarnings(t, aliasStderr)
-
-	var canonicalOut struct {
-		Data struct {
-			ID string `json:"id"`
-		} `json:"data"`
-	}
-	if err := json.Unmarshal([]byte(canonicalStdout), &canonicalOut); err != nil {
-		t.Fatalf("unmarshal canonical output: %v\nstdout: %s", err, canonicalStdout)
-	}
-
-	var aliasOut struct {
-		Data struct {
-			ID string `json:"id"`
-		} `json:"data"`
-	}
-	if err := json.Unmarshal([]byte(aliasStdout), &aliasOut); err != nil {
-		t.Fatalf("unmarshal alias output: %v\nstdout: %s", err, aliasStdout)
-	}
-
-	if canonicalOut.Data.ID != "build-latest" || aliasOut.Data.ID != "build-latest" {
-		t.Fatalf("expected both canonical and alias to resolve build-latest, canonical=%q alias=%q", canonicalOut.Data.ID, aliasOut.Data.ID)
-	}
-}
-
 func TestBuildsHelpShowsNextBuildNumberAndHidesLatestAlias(t *testing.T) {
 	usage := usageForCommand(t, "builds")
 	if !strings.Contains(usage, "\n  next-build-number") {
