+++ internal/cli/release/run.go
@@ -4,14 +4,11 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
-	"flag"
 	"fmt"
 	"os"
 	"path/filepath"
 	"strings"
 	"time"
-
-	"github.com/peterbourgon/ff/v3/ffcli"
 
 	"github.com/rudrankriyam/App-Store-Connect-CLI/internal/asc"
 	"github.com/rudrankriyam/App-Store-Connect-CLI/internal/cli/metadata"
@@ -22,15 +19,14 @@ import (
 )
 
 const (
-	stepEnsureVersion            = "ensure_version"
-	stepApplyMetadata            = "apply_metadata"
-	stepAttachBuild              = "attach_build"
-	stepValidateReadiness        = "validate_readiness"
-	stepSubmitReview             = "submit_review"
-	releaseModeRun               = "run"
-	releaseModeStage             = "stage"
-	releaseRunTimeout            = 30 * time.Minute
-	releaseRunDeprecationWarning = "Warning: `asc release run` is deprecated. Use `asc release stage`, then `asc review submissions-create` / `asc review items-add` / `asc review submissions-submit` for metadata workflows, or `asc publish appstore --submit` when local metadata is already synced."
+	stepEnsureVersion     = "ensure_version"
+	stepApplyMetadata     = "apply_metadata"
+	stepAttachBuild       = "attach_build"
+	stepValidateReadiness = "validate_readiness"
+	stepSubmitReview      = "submit_review"
+	releaseModeRun        = "run"
+	releaseModeStage      = "stage"
+	releaseRunTimeout     = 30 * time.Minute
 )
 
 var (
@@ -109,126 +105,6 @@ type stepOutcome struct {
 	Persist      bool
 	ResolvedID   string
 	SubmissionID string
-}
-
-// ReleaseRunCommand runs the end-to-end release orchestration flow.
-func ReleaseRunCommand() *ffcli.Command {
-	fs := flag.NewFlagSet("release run", flag.ExitOnError)
-
-	appID := fs.String("app", "", "App Store Connect app ID (or ASC_APP_ID)")
-	version := fs.String("version", "", "App Store version string (required)")
-	buildID := fs.String("build", "", "Build ID to attach (required)")
-	metadataDir := fs.String("metadata-dir", "", "Metadata directory to apply (required)")
-	platform := fs.String("platform", "IOS", "Platform: IOS, MAC_OS, TV_OS, VISION_OS")
-	timeout := fs.Duration("timeout", releaseRunTimeout, "Maximum time to run the release pipeline")
-	dryRun := fs.Bool("dry-run", false, "Preview deterministic plan without mutations")
-	confirm := fs.Bool("confirm", false, "Confirm release mutations (required unless --dry-run)")
-	strictValidate := fs.Bool("strict-validate", false, "Treat readiness warnings as blocking")
-	checkpointFile := fs.String("checkpoint-file", "", "Checkpoint path for resumable runs")
-	output := shared.BindOutputFlags(fs)
-
-	return &ffcli.Command{
-		Name:       "run",
-		ShortUsage: "asc release run --app \"APP_ID\" --version \"2.4.0\" --build \"BUILD_ID\" --metadata-dir \"./metadata/version/2.4.0\" [flags]",
-		ShortHelp:  "DEPRECATED: use `asc release stage` + `asc review submissions-*`, or `asc publish appstore --submit`.",
-		LongHelp: `DEPRECATED: use ` + "`asc release stage`" + ` + ` + "`asc review submissions-*`" + ` when you need the metadata workflow, or ` + "`asc publish appstore --submit`" + ` when local metadata is already synced.
-
-Deprecated compatibility pipeline that still runs the old single-command
-release flow:
-1. Ensure/create version
-2. Apply metadata/localizations
-3. Attach selected build
-4. Run readiness checks
-5. Submit for review
-
-	Prefer:
-	  - ` + "`asc release stage`" + ` + ` + "`asc review submissions-create`" + ` / ` + "`asc review items-add`" + ` / ` + "`asc review submissions-submit`" + ` when you need the old metadata-dir staging + submit workflow
-	  - ` + "`asc publish appstore --submit`" + ` for the canonical high-level App Store upload + submit flow when local metadata is already applied
-
-Supports dry-run planning, step-level structured output, and checkpointed resume.
-
-Examples:
-  asc release run --app "APP_ID" --version "2.4.0" --build "BUILD_ID" --metadata-dir "./metadata/version/2.4.0" --dry-run
-  asc release run --app "APP_ID" --version "2.4.0" --build "BUILD_ID" --metadata-dir "./metadata/version/2.4.0" --confirm`,
-		FlagSet:   fs,
-		UsageFunc: shared.DefaultUsageFunc,
-		Exec: func(ctx context.Context, args []string) error {
-			fmt.Fprintln(os.Stderr, releaseRunDeprecationWarning)
-			if len(args) > 0 {
-				return shared.UsageError("release run does not accept positional arguments")
-			}
-			if !*dryRun && !*confirm {
-				return shared.UsageError("--confirm is required unless --dry-run is set")
-			}
-
-			resolvedAppID := shared.ResolveAppID(*appID)
-			if strings.TrimSpace(resolvedAppID) == "" {
-				return shared.UsageError("--app is required (or set ASC_APP_ID)")
-			}
-			trimmedVersion := strings.TrimSpace(*version)
-			if trimmedVersion == "" {
-				return shared.UsageError("--version is required")
-			}
-			trimmedBuildID := strings.TrimSpace(*buildID)
-			if trimmedBuildID == "" {
-				return shared.UsageError("--build is required")
-			}
-			trimmedMetadataDir := strings.TrimSpace(*metadataDir)
-			if trimmedMetadataDir == "" {
-				return shared.UsageError("--metadata-dir is required")
-			}
-
-			normalizedPlatform, err := shared.NormalizeAppStoreVersionPlatform(*platform)
-			if err != nil {
-				return shared.UsageError(err.Error())
-			}
-			if *timeout <= 0 {
-				return shared.UsageError("--timeout must be greater than 0")
-			}
-
-			checkpointPath := strings.TrimSpace(*checkpointFile)
-			if checkpointPath == "" {
-				checkpointPath = defaultCheckpointPath(resolvedAppID, trimmedVersion, trimmedBuildID, normalizedPlatform)
-			}
-			absCheckpointPath, err := filepath.Abs(checkpointPath)
-			if err != nil {
-				return fmt.Errorf("release run: resolve checkpoint path: %w", err)
-			}
-
-			result, runErr := executeRun(ctx, runOptions{
-				AppID:          resolvedAppID,
-				Version:        trimmedVersion,
-				BuildID:        trimmedBuildID,
-				MetadataDir:    trimmedMetadataDir,
-				Platform:       normalizedPlatform,
-				Timeout:        *timeout,
-				DryRun:         *dryRun,
-				Confirm:        *confirm,
-				StrictValidate: *strictValidate,
-				CheckpointFile: absCheckpointPath,
-			})
-
-			if printErr := shared.PrintOutput(result, *output.Output, *output.Pretty); printErr != nil {
-				return printErr
-			}
-			if runErr != nil {
-				return shared.NewReportedError(runErr)
-			}
-			return nil
-		},
-	}
-}
-
-func RemovedReleaseRunCommand() *ffcli.Command {
-	cmd := ReleaseRunCommand()
-	cmd.ShortHelp = "DEPRECATED: removed; use `asc release stage`."
-	cmd.LongHelp = "Removed legacy command. Use `asc release stage` instead."
-	cmd.UsageFunc = shared.DeprecatedUsageFunc
-	cmd.Exec = func(ctx context.Context, args []string) error {
-		fmt.Fprintln(os.Stderr, "Error: `asc release run` was removed. Use `asc release stage` instead.")
-		return flag.ErrHelp
-	}
-	return cmd
 }
 
 func executeRun(ctx context.Context, opts runOptions) (runResult, error) {
