+++ internal/cli/submit/submit.go
@@ -38,216 +38,9 @@ Use:
 		Subcommands: []*ffcli.Command{
 			SubmitStatusCommand(),
 			SubmitCancelCommand(),
-			RemovedSubmitCreateCommand(),
-			RemovedSubmitPreflightCommand(),
 		},
 		Exec: func(ctx context.Context, args []string) error {
 			return flag.ErrHelp
-		},
-	}
-}
-
-func RemovedSubmitCreateCommand() *ffcli.Command {
-	cmd := SubmitCreateCommand()
-	cmd.ShortHelp = "DEPRECATED: removed; use `asc publish appstore --submit` or `asc review submit`."
-	cmd.LongHelp = "Removed legacy command. Use `asc publish appstore --submit` for the canonical high-level flow, or `asc review submit` when the build is already uploaded and the version is already prepared."
-	cmd.UsageFunc = shared.DeprecatedUsageFunc
-	cmd.Exec = func(ctx context.Context, args []string) error {
-		fmt.Fprintln(os.Stderr, "Error: `asc submit create` was removed. Use `asc publish appstore --submit` or `asc review submit` instead.")
-		return flag.ErrHelp
-	}
-	return cmd
-}
-
-func SubmitCreateCommand() *ffcli.Command {
-	fs := flag.NewFlagSet("submit create", flag.ExitOnError)
-
-	appID := fs.String("app", "", "App Store Connect app ID (or ASC_APP_ID)")
-	version := fs.String("version", "", "App Store version string")
-	versionID := fs.String("version-id", "", "App Store version ID")
-	buildID := fs.String("build", "", "Build ID to attach")
-	platform := fs.String("platform", "IOS", "Platform: IOS, MAC_OS, TV_OS, VISION_OS")
-	confirm := fs.Bool("confirm", false, "Confirm submission (required)")
-	output := shared.BindOutputFlags(fs)
-
-	return &ffcli.Command{
-		Name:       "create",
-		ShortUsage: "asc submit create [flags]",
-		ShortHelp:  "DEPRECATED: use `asc review submit`.",
-		LongHelp: `Deprecated compatibility path for lower-level direct submission.
-
-For already-uploaded builds, use:
-  - ` + "`asc review submit --app \"APP_ID\" --version-id \"VERSION_ID\" --build \"BUILD_ID\" --confirm`" + `
-
-Use ` + "`asc publish appstore --submit`" + ` only when you are starting from
-an IPA upload or local build.
-
-Keep ` + "`asc submit create`" + ` only for older automation that already prepared
-the version and just needs the final review-submission step. For newly scripted
-direct submission on an already-prepared version, prefer the
-nondeprecated command above instead of extending this alias.`,
-		FlagSet:   fs,
-		UsageFunc: shared.DeprecatedUsageFunc,
-		Exec: func(ctx context.Context, args []string) error {
-			fmt.Fprintln(os.Stderr, "Warning: `asc submit create` is deprecated. Use `asc review submit` for already-uploaded builds, or `asc publish appstore --submit` when starting from an IPA.")
-			if !*confirm {
-				fmt.Fprintln(os.Stderr, "Error: --confirm is required to submit for review")
-				return flag.ErrHelp
-			}
-			if strings.TrimSpace(*buildID) == "" {
-				fmt.Fprintln(os.Stderr, "Error: --build is required")
-				return flag.ErrHelp
-			}
-			if strings.TrimSpace(*version) == "" && strings.TrimSpace(*versionID) == "" {
-				fmt.Fprintln(os.Stderr, "Error: --version or --version-id is required")
-				return flag.ErrHelp
-			}
-			if strings.TrimSpace(*version) != "" && strings.TrimSpace(*versionID) != "" {
-				return shared.UsageError("--version and --version-id are mutually exclusive")
-			}
-
-			resolvedAppID := shared.ResolveAppID(*appID)
-			if resolvedAppID == "" {
-				fmt.Fprintln(os.Stderr, "Error: --app is required (or set ASC_APP_ID)")
-				return flag.ErrHelp
-			}
-
-			normalizedPlatform, err := shared.NormalizeAppStoreVersionPlatform(*platform)
-			if err != nil {
-				return shared.UsageError(err.Error())
-			}
-
-			client, err := shared.GetASCClient()
-			if err != nil {
-				return fmt.Errorf("submit create: %w", err)
-			}
-
-			resolvedVersionID := strings.TrimSpace(*versionID)
-			effectivePlatform := normalizedPlatform
-			if resolvedVersionID == "" {
-				resolveCtx, resolveCancel := shared.ContextWithTimeout(ctx)
-				resolvedVersionID, err = shared.ResolveAppStoreVersionID(resolveCtx, client, resolvedAppID, strings.TrimSpace(*version), normalizedPlatform)
-				resolveCancel()
-				if err != nil {
-					return fmt.Errorf("submit create: %w", err)
-				}
-			} else {
-				versionCtx, versionCancel := shared.ContextWithTimeout(ctx)
-				versionResp, versionErr := client.GetAppStoreVersion(versionCtx, resolvedVersionID)
-				versionCancel()
-				if versionErr != nil {
-					return fmt.Errorf("submit create: failed to fetch version %q: %w", resolvedVersionID, versionErr)
-				}
-
-				effectivePlatform, err = shared.NormalizeAppStoreVersionPlatform(string(versionResp.Data.Attributes.Platform))
-				if err != nil {
-					return fmt.Errorf("submit create: version %q returned unsupported platform %q", resolvedVersionID, string(versionResp.Data.Attributes.Platform))
-				}
-			}
-
-			if err := runSubmitCreateLocalizationPreflight(ctx, client, resolvedAppID, resolvedVersionID, effectivePlatform, 0); err != nil {
-				return err
-			}
-
-			resolvedBuildID := strings.TrimSpace(*buildID)
-			if err := runSubmitCreateReadinessPreflight(ctx, client, resolvedAppID, resolvedVersionID, effectivePlatform, resolvedBuildID); err != nil {
-				return err
-			}
-
-			// Attach the build only after blocking preflight checks pass so a
-			// failed submit create does not mutate the version unnecessarily.
-			attachCtx, attachCancel := shared.ContextWithTimeout(ctx)
-			if err := client.AttachBuildToVersion(attachCtx, resolvedVersionID, resolvedBuildID); err != nil {
-				attachCancel()
-				return fmt.Errorf("submit create: failed to attach build: %w", err)
-			}
-			attachCancel()
-
-			runSubmitCreateSubscriptionPreflight(ctx, client, resolvedAppID, 0)
-
-			preparationCtx, preparationCancel := shared.ContextWithTimeout(ctx)
-			preparedSubmission := prepareReviewSubmissionForCreate(preparationCtx, client, resolvedAppID, effectivePlatform, resolvedVersionID, nil)
-			preparationCancel()
-
-			requestCtx, cancel := shared.ContextWithTimeout(ctx)
-			defer func() {
-				if cancel != nil {
-					cancel()
-				}
-			}()
-
-			var createdSubmissionID string
-			submissionIDToSubmit := strings.TrimSpace(preparedSubmission.reuseSubmissionID)
-
-			if submissionIDToSubmit == "" {
-				// Use the new reviewSubmissions API (the old appStoreVersionSubmissions is deprecated)
-				// Step 1: Create review submission for the app
-				reviewSubmission, err := client.CreateReviewSubmission(requestCtx, resolvedAppID, asc.Platform(effectivePlatform))
-				if err != nil {
-					return fmt.Errorf("submit create: failed to create review submission: %w", err)
-				}
-				createdSubmissionID = strings.TrimSpace(reviewSubmission.Data.ID)
-				submissionIDToSubmit = createdSubmissionID
-			}
-
-			if !preparedSubmission.reuseSubmissionHasVersion {
-				// Step 2: Add the app store version as a submission item.
-				// If the version is already in another submission, recover by
-				// submitting that existing submission instead. If the conflicting
-				// submission is one we just canceled as stale, retry the add until
-				// App Store Connect finishes detaching the version.
-				submissionIDToSubmit, err = addVersionToSubmissionOrRecover(
-					requestCtx,
-					client,
-					submissionIDToSubmit,
-					resolvedVersionID,
-					preparedSubmission.canceledSubmissionIDs,
-					func(message string) {
-						fmt.Fprintln(os.Stderr, message)
-					},
-				)
-				if err != nil {
-					if createdSubmissionID != "" {
-						cleanupEmptyReviewSubmission(requestCtx, client, createdSubmissionID, nil)
-					}
-					printSubmissionErrorHints(err, submissionErrorHintContext{
-						AppID:         resolvedAppID,
-						Platform:      effectivePlatform,
-						VersionID:     resolvedVersionID,
-						VersionString: strings.TrimSpace(*version),
-					})
-					return fmt.Errorf("submit create: failed to add version to submission: %w", err)
-				}
-			}
-			if createdSubmissionID != "" && submissionIDToSubmit != createdSubmissionID {
-				cleanupEmptyReviewSubmission(requestCtx, client, createdSubmissionID, nil)
-			}
-
-			// Step 3: Submit for review
-			submitResp, err := client.SubmitReviewSubmission(requestCtx, submissionIDToSubmit)
-			if err != nil {
-				printSubmissionErrorHints(err, submissionErrorHintContext{
-					AppID:         resolvedAppID,
-					Platform:      effectivePlatform,
-					VersionID:     resolvedVersionID,
-					VersionString: strings.TrimSpace(*version),
-				})
-				return fmt.Errorf("submit create: failed to submit for review: %w", err)
-			}
-
-			submittedDate := submitResp.Data.Attributes.SubmittedDate
-			var createdDatePtr *string
-			if submittedDate != "" {
-				createdDatePtr = &submittedDate
-			}
-			result := &asc.AppStoreVersionSubmissionCreateResult{
-				SubmissionID: submitResp.Data.ID,
-				VersionID:    resolvedVersionID,
-				BuildID:      resolvedBuildID,
-				CreatedDate:  createdDatePtr,
-			}
-
-			return shared.PrintOutput(result, *output.Output, *output.Pretty)
 		},
 	}
 }
